Move X035 fallback detection into facts
